refactor: remove HederaAliasManager and introduce auto creation hierarchy
