--- hedera-node/hedera-mono-service/src/main/java/com/hedera/node/app/service/mono/ledger/accounts/HederaAliasManager.java
@@ -1,32 +1,0 @@
-/*
- * Copyright (C) 2022 Hedera Hashgraph, LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.hedera.node.app.service.mono.ledger.accounts;
-
-import com.google.protobuf.ByteString;
-import com.hedera.node.app.service.mono.legacy.core.jproto.JKey;
-import com.hedera.node.app.service.mono.utils.EntityNum;
-import edu.umd.cs.findbugs.annotations.Nullable;
-
-public interface HederaAliasManager {
-
-    void link(final ByteString alias, final EntityNum num);
-
-    boolean maybeLinkEvmAddress(@Nullable final JKey key, final EntityNum num);
-
-    void unlink(final ByteString alias);
-
-    void forgetEvmAddress(final ByteString alias);
-}
