+++ hedera-node/hedera-mono-service/src/main/java/com/hedera/node/app/service/mono/store/contracts/precompile/InfrastructureFactory.java
@@ -56,9 +56,10 @@ import com.hedera.node.app.service.mono.store.contracts.precompile.proxy.ViewExe
 import com.hedera.node.app.service.mono.store.contracts.precompile.proxy.ViewGasCalculator;
 import com.hedera.node.app.service.mono.store.models.NftId;
 import com.hedera.node.app.service.mono.store.tokens.HederaTokenStore;
+import com.hedera.node.app.service.mono.txns.crypto.AbstractAutoCreationLogic;
 import com.hedera.node.app.service.mono.txns.crypto.ApproveAllowanceLogic;
-import com.hedera.node.app.service.mono.txns.crypto.AutoCreationLogic;
 import com.hedera.node.app.service.mono.txns.crypto.DeleteAllowanceLogic;
+import com.hedera.node.app.service.mono.txns.crypto.EvmAutoCreationLogic;
 import com.hedera.node.app.service.mono.txns.crypto.validators.ApproveAllowanceChecks;
 import com.hedera.node.app.service.mono.txns.crypto.validators.DeleteAllowanceChecks;
 import com.hedera.node.app.service.mono.txns.customfees.CustomFeeSchedules;
@@ -346,19 +347,18 @@ public class InfrastructureFactory {
                 sigImpactHistorian);
     }
 
-    public AutoCreationLogic newAutoCreationLogicScopedTo(
+    public AbstractAutoCreationLogic newAutoCreationLogicScopedTo(
             final HederaStackedWorldStateUpdater updater) {
         final var autoCreationLogic =
-                new AutoCreationLogic(
+                new EvmAutoCreationLogic(
                         usageLimits,
                         syntheticTxnFactory,
                         entityCreator,
                         ids,
-                        updater.aliases(),
-                        sigImpactHistorian,
                         () -> view,
                         txnCtx,
-                        dynamicProperties);
+                        dynamicProperties,
+                        updater.aliases());
         autoCreationLogic.setFeeCalculator(feeCalculator.get());
         return autoCreationLogic;
     }
